DeleteCommand refactored (aurusov/raox#334)
