+++ ru.bmstu.rk9.rao.ui/src/ru/bmstu/rk9/rao/ui/process/ProcessDeletePolicy.java
@@ -5,12 +5,12 @@ import org.eclipse.gef.editpolicies.ComponentEditPolicy;
 import org.eclipse.gef.requests.GroupRequest;
 
 import ru.bmstu.rk9.rao.ui.gef.Node;
-import ru.bmstu.rk9.rao.ui.process.commands.DeleteCommand;
+import ru.bmstu.rk9.rao.ui.process.commands.DeleteBlockCommand;
 
 public class ProcessDeletePolicy extends ComponentEditPolicy {
 
 	@Override
 	protected Command createDeleteCommand(GroupRequest deleteRequest) {
-		return new DeleteCommand((Node) getHost().getModel(), (Node) getHost().getParent().getModel());
+		return new DeleteBlockCommand((Node) getHost().getModel(), (Node) getHost().getParent().getModel());
 	}
 }
