+++ ru.bmstu.rk9.rao.ui/src/ru/bmstu/rk9/rao/ui/process/commands/DeleteBlockCommand.java
@@ -3,35 +3,25 @@ package ru.bmstu.rk9.rao.ui.process.commands;
 import java.util.Iterator;
 import java.util.List;
 
-import org.eclipse.gef.commands.Command;
-
 import ru.bmstu.rk9.rao.ui.gef.Node;
+import ru.bmstu.rk9.rao.ui.gef.commands.CommandDelete;
 import ru.bmstu.rk9.rao.ui.process.connection.Connection;
 import ru.bmstu.rk9.rao.ui.process.node.BlockNode;
 
-public class DeleteCommand extends Command {
+public class DeleteBlockCommand extends CommandDelete {
 
-	private Node node;
-	private Node parentNode;
-
-	public DeleteCommand(Node node, Node parentNode) {
-		this.node = node;
-		this.parentNode = parentNode;
+	public DeleteBlockCommand(Node node, Node parentNode) {
+		super(node, parentNode);
 	}
 
 	@Override
 	public void execute() {
-		parentNode.removeChild(node);
+		super.execute();
 		if (node instanceof BlockNode) {
 			BlockNode blockNode = (BlockNode) node;
 			disconnect(blockNode.getSourceConnections());
 			disconnect(blockNode.getTargetConnections());
 		}
-	}
-
-	@Override
-	public void undo() {
-		parentNode.addChild(node);
 	}
 
 	private final void disconnect(List<Connection> connections) {
