+++ ru.bmstu.rk9.rao.ui/src/ru/bmstu/rk9/rao/ui/process/label/LabelEditPart.java
@@ -3,11 +3,9 @@ package ru.bmstu.rk9.rao.ui.process.label;
 import java.beans.PropertyChangeEvent;
 
 import org.eclipse.draw2d.IFigure;
-import org.eclipse.gef.EditPolicy;
 import org.eclipse.swt.graphics.Color;
 
 import ru.bmstu.rk9.rao.ui.gef.EditPart;
-import ru.bmstu.rk9.rao.ui.process.ProcessDeletePolicy;
 
 public class LabelEditPart extends EditPart {
 
@@ -37,10 +35,5 @@ public class LabelEditPart extends EditPart {
 				|| evt.getPropertyName().equals(LabelNode.PROPERTY_BACKGROUND_COLOR)) {
 			refreshVisuals();
 		}
-	}
-
-	@Override
-	protected void createEditPolicies() {
-		installEditPolicy(EditPolicy.COMPONENT_ROLE, new ProcessDeletePolicy());
 	}
 }
